requirements.txt
pip==21.3.1
setuptools==59.6.0
wheel==0.37.1
git+https://github.com/aclements/biblib.git#egg=biblib
jellyfish
pycbc
matplotlib
healpy
ligo.skymap
pyqt5
mayavi
